# tsdb: use dennwc/varint to speed up WAL decoding